Use project scaffold
