--- examples/with_openai/requirements.txt
@@ -1,1 +1,0 @@
-langchain==0.1.11
