+++ examples/with_openai/with_openai/constants.py
@@ -1,0 +1,6 @@
+SEARCH_INDEX_FILE = "search_index.pickle"
+
+SUMMARY_TEMPLATE = """
+Content: {content}
+Source: {source}
+"""
