+++ examples/with_openai/with_openai/partitions.py
@@ -1,0 +1,3 @@
+from dagster import StaticPartitionsDefinition
+
+docs_partitions_def = StaticPartitionsDefinition(["guides", "integrations"])
